got counter publish subscribe working

+++ include/output_nodes/dataCache.hpp
@@ -5,12 +5,12 @@ public:
     dataCache(int size) : size(size), buffer(size), head(0), tail(0), count(0) {}
     bool isFull();
     bool isEmpty();
-    void enqueue(double data);
-    double newestValue();
+    void enqueue(int data);
+    int newestValue();
 
 private:
     int size;
-    std::vector<double> buffer = {};
+    std::vector<int> buffer = {}; 
     int head;
     int tail;
     int count;
